# lib/data/network/swagger/money_mate_api.json
{"openapi":"3.0.0","paths":{"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User successfully registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"summary":"Register a new user","tags":["Authentication"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User successfully logged in","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"summary":"Login with email and password","tags":["Authentication"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"Token successfully refreshed","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"summary":"Refresh access token","tags":["Authentication"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"Successfully logged out","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Logout user","tags":["Authentication"]}},"/api/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"Reset link sent if email exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordResponseDto"}}}}},"summary":"Send password reset email","tags":["Authentication"]}},"/api/users/me":{"get":{"operationId":"UsersController_getProfile","parameters":[],"responses":{"200":{"description":"Profile successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get current user's profile","tags":["Users"]},"delete":{"operationId":"UsersController_deleteAccount","parameters":[],"responses":{"200":{"description":"Account successfully deleted"}},"security":[{"bearer":[]}],"summary":"Delete current user account","tags":["Users"]}},"/api/users/me/settings":{"get":{"operationId":"UsersController_getSettings","parameters":[],"responses":{"200":{"description":"Settings successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSettingsResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get current user's settings","tags":["Users"]},"patch":{"operationId":"UsersController_updateSettings","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSettingsDto"}}}},"responses":{"200":{"description":"Settings successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSettingsResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update current user's settings","tags":["Users"]}},"/api/categories":{"get":{"operationId":"CategoriesController_findAll","parameters":[{"name":"isIncome","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Categories successfully retrieved","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponseDto"}}}}}},"security":[{"bearer":[]}],"summary":"Get all categories","tags":["Categories"]},"post":{"operationId":"CategoriesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"Category successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Create a new category","tags":["Categories"]},"delete":{"operationId":"CategoriesController_removeAll","parameters":[{"name":"isIncome","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"All categories successfully deleted"}},"security":[{"bearer":[]}],"summary":"Delete all categories for user","tags":["Categories"]}},"/api/categories/{id}":{"get":{"operationId":"CategoriesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Category successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get category by ID","tags":["Categories"]},"put":{"operationId":"CategoriesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":"Category successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update category","tags":["Categories"]},"delete":{"operationId":"CategoriesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Category successfully deleted"}},"security":[{"bearer":[]}],"summary":"Delete category","tags":["Categories"]}},"/api/categories/{id}/limit":{"patch":{"operationId":"CategoriesController_updateLimit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLimitDto"}}}},"responses":{"200":{"description":"Category limit successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update category limit","tags":["Categories"]}},"/api/categories/restore-all-limits":{"post":{"operationId":"CategoriesController_restoreAllLimits","parameters":[],"responses":{"200":{"description":"All category limits successfully restored"}},"security":[{"bearer":[]}],"summary":"Restore all category limits","tags":["Categories"]}},"/api/transactions":{"get":{"operationId":"TransactionsController_findAll","parameters":[{"name":"monthYear","required":true,"in":"query","schema":{"type":"string"}},{"name":"year","required":true,"in":"query","schema":{"type":"string"}},{"name":"isIncome","required":true,"in":"query","schema":{"type":"string"}},{"name":"catId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Transactions successfully retrieved","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponseDto"}}}}}},"security":[{"bearer":[]}],"summary":"Get all transactions with filters","tags":["Transactions"]},"post":{"operationId":"TransactionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTransactionDto"}}}},"responses":{"201":{"description":"Transaction successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Create a new transaction","tags":["Transactions"]},"delete":{"operationId":"TransactionsController_removeAll","parameters":[],"responses":{"200":{"description":"All transactions successfully deleted"}},"security":[{"bearer":[]}],"summary":"Delete all transactions for user","tags":["Transactions"]}},"/api/transactions/{id}":{"put":{"operationId":"TransactionsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTransactionDto"}}}},"responses":{"200":{"description":"Transaction successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update transaction","tags":["Transactions"]},"delete":{"operationId":"TransactionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Transaction successfully deleted"}},"security":[{"bearer":[]}],"summary":"Delete transaction","tags":["Transactions"]}},"/api/transactions/search":{"get":{"operationId":"TransactionsController_search","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Search results retrieved","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponseDto"}}}}}},"security":[{"bearer":[]}],"summary":"Search transactions","tags":["Transactions"]}},"/api/transactions/yearly":{"get":{"operationId":"TransactionsController_getYearlyStats","parameters":[{"name":"year","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Yearly stats retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/YearlyStatsResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get yearly statistics","tags":["Transactions"]}},"/api/schedules":{"get":{"operationId":"SchedulesController_findAll","parameters":[],"responses":{"200":{"description":"Schedules successfully retrieved","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}}},"security":[{"bearer":[]}],"summary":"Get all schedules","tags":["Schedules"]},"post":{"operationId":"SchedulesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateScheduleDto"}}}},"responses":{"201":{"description":"Schedule successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Create a new schedule","tags":["Schedules"]},"delete":{"operationId":"SchedulesController_removeAll","parameters":[],"responses":{"200":{"description":"All schedules successfully deleted"}},"security":[{"bearer":[]}],"summary":"Delete all schedules for user","tags":["Schedules"]}},"/api/schedules/{id}":{"delete":{"operationId":"SchedulesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Schedule successfully deleted"}},"security":[{"bearer":[]}],"summary":"Delete schedule by ID","tags":["Schedules"]}}},"info":{"title":"API Money Mate","description":"The API Money Mate description","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"Authorization","description":"Enter JWT token","in":"header"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"password123","minLength":6},"confirmPassword":{"type":"string","example":"password123","minLength":6}},"required":["email","password","confirmPassword"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"email":{"type":"string","example":"user@example.com"},"language":{"type":"string","example":"en"},"isDark":{"type":"boolean","example":false},"isLock":{"type":"boolean","example":false}},"required":["id","email","language","isDark","isLock"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserResponseDto"}},"required":["accessToken","refreshToken","user"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"password123","minLength":6}},"required":["email","password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"LogoutResponseDto":{"type":"object","properties":{"message":{"type":"string","example":"Logged out successfully"}},"required":["message"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"}},"required":["email"]},"ForgotPasswordResponseDto":{"type":"object","properties":{"message":{"type":"string","example":"If email exists, password reset link will be sent"}},"required":["message"]},"UserSettingsResponseDto":{"type":"object","properties":{"language":{"type":"string","example":"en"},"isDark":{"type":"boolean","example":false},"isLock":{"type":"boolean","example":false}},"required":["language","isDark","isLock"]},"UpdateSettingsDto":{"type":"object","properties":{"language":{"type":"string","example":"en"},"isDark":{"type":"boolean","example":false},"isLock":{"type":"boolean","example":false}}},"CategoryResponseDto":{"type":"object","properties":{"id":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"icon":{"type":"string","example":"🍔"},"name":{"type":"string","example":"Food"},"isIncome":{"type":"boolean","example":false},"limit":{"type":"number","example":1000000},"userId":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"}},"required":["id","icon","name","isIncome","userId"]},"CreateCategoryDto":{"type":"object","properties":{"icon":{"type":"string","example":"🍔"},"name":{"type":"string","example":"Food"},"isIncome":{"type":"boolean","example":false},"limit":{"type":"number","example":1000000}},"required":["icon","name","isIncome"]},"UpdateCategoryDto":{"type":"object","properties":{"icon":{"type":"string","example":"🍔"},"name":{"type":"string","example":"Food"},"isIncome":{"type":"boolean","example":false},"limit":{"type":"number","example":1000000}}},"UpdateLimitDto":{"type":"object","properties":{"limit":{"type":"number","example":1500000}},"required":["limit"]},"TransactionResponseDto":{"type":"object","properties":{"id":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"date":{"type":"string","example":"2024-03-20T00:00:00.000Z"},"description":{"type":"string","example":"Buy coffee"},"money":{"type":"number","example":50000},"catId":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"isIncome":{"type":"boolean","example":false},"userId":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"category":{"$ref":"#/components/schemas/CategoryResponseDto"}},"required":["id","date","money","catId","isIncome","userId"]},"CreateTransactionDto":{"type":"object","properties":{"date":{"type":"string","example":"2024-03-20"},"description":{"type":"string","example":"Buy coffee"},"money":{"type":"number","example":50000},"catId":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"isIncome":{"type":"boolean","example":false}},"required":["date","money","catId","isIncome"]},"UpdateTransactionDto":{"type":"object","properties":{"date":{"type":"string","example":"2024-03-20"},"description":{"type":"string","example":"Buy coffee"},"money":{"type":"number","example":50000},"catId":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"isIncome":{"type":"boolean","example":false}}},"TransactionSummaryDto":{"type":"object","properties":{"income":{"type":"number","example":1000000},"expense":{"type":"number","example":500000},"balance":{"type":"number","example":500000}},"required":["income","expense","balance"]},"YearlyStatsResponseDto":{"type":"object","properties":{"year":{"type":"string","example":"2024"},"transactions":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponseDto"}},"summary":{"$ref":"#/components/schemas/TransactionSummaryDto"}},"required":["year","transactions","summary"]},"ScheduleResponseDto":{"type":"object","properties":{"id":{"type":"number","example":1},"date":{"type":"string","example":"2024-03-20"},"description":{"type":"string","example":"Monthly rent"},"money":{"type":"number","example":5000000},"catId":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"icon":{"type":"string","example":"🏠"},"name":{"type":"string","example":"Rent"},"isIncome":{"type":"boolean","example":false},"option":{"type":"string","example":"MONTHLY","enum":["DAILY","WEEKLY","MONTHLY","YEARLY"]},"userId":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"}},"required":["id","date","money","catId","icon","name","isIncome","option","userId"]},"CreateScheduleDto":{"type":"object","properties":{"date":{"type":"string","example":"2024-03-20"},"description":{"type":"string","example":"Monthly rent"},"money":{"type":"number","example":5000000},"catId":{"type":"string","example":"123e4567-e89b-12d3-a456-426614174000"},"icon":{"type":"string","example":"🏠"},"name":{"type":"string","example":"Rent"},"isIncome":{"type":"boolean","example":false},"option":{"type":"string","enum":["Never","Daily","Weekly","Monthly","Yearly"],"example":"Monthly"}},"required":["date","money","catId","icon","name","isIncome","option"]}}},"externalDocs":{"description":"JSON DOCUMENTS","url":"/api/docs-json"}}