# moneymate-nestjs-backend/src/auth/dto/auth-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class UserResponseDto {
  @ApiProperty({ example: '123e4567-e89b-12d3-a456-426614174000' })
  id: string;

  @ApiProperty({ example: 'user@example.com' })
  email: string;

  @ApiProperty({ example: 'en' })
  language: string;

  @ApiProperty({ example: false })
  isDark: boolean;

  @ApiProperty({ example: false })
  isLock: boolean;
}

export class AuthResponseDto {
  @ApiProperty()
  accessToken: string;

  @ApiProperty()
  refreshToken: string;

  @ApiProperty({ type: UserResponseDto })
  user: UserResponseDto;
}

export class LogoutResponseDto {
  @ApiProperty({ example: 'Logged out successfully' })
  message: string;
}

export class ForgotPasswordResponseDto {
  @ApiProperty({ example: 'If email exists, password reset link will be sent' })
  message: string;
}